refactor: update ext protocol registration
